+++ docs/examples/clientside_data/rowTransaction.py
@@ -8,7 +8,7 @@ app = Dash(__name__)
 rowData = [
     {"make": "Toyota", "model": "Celica", "price": 35000},
     {"make": "Ford", "model": "Mondeo", "price": 32000},
-    {"make": "Porsche", "model": "Boxter", "price": 72000},
+    {"make": "Porsche", "model": "Boxster", "price": 72000},
 ]
 
 columnDefs = [