+++ tests/test_default_styles.py
@@ -15,7 +15,7 @@ def test_ds001_default_styles(dash_duo):
     rowData = [
         {"make": "Toyota", "model": "Celica", "price": 35000},
         {"make": "Ford", "model": "Mondeo", "price": 32000},
-        {"make": "Porsche", "model": "Boxter", "price": 72000},
+        {"make": "Porsche", "model": "Boxster", "price": 72000},
     ]
 
     grid_with_default_cell_styles = html.Div(