+++ tests/test_row_menu.py
@@ -40,7 +40,7 @@ def test_rm001_row_menu(dash_duo):
                     },
                     {
                         "make": "Porsche",
-                        "model": "Boxter",
+                        "model": "Boxster",
                         "price": 72000,
                         "menu": [
                             {"label": "Option 7", "value": 7},